moving JMX ssl distributed test to jmx package
