+++ test/distributed/org/apache/cassandra/distributed/test/AbstractEncryptionOptionsImpl.java
@@ -70,7 +70,7 @@ public class AbstractEncryptionOptionsImpl extends TestBaseImpl
     final static String validTrustStorePassword = TlsTestUtils.SERVER_TRUSTSTORE_PASSWORD;
 
     // Base configuration map for a valid keystore that can be opened
-    final static Map<String,Object> validKeystore = ImmutableMap.of("keystore", validKeyStorePath,
+    protected final static Map<String,Object> validKeystore = ImmutableMap.of("keystore", validKeyStorePath,
                                                                     "keystore_password", validKeyStorePassword,
                                                                     "truststore", validTrustStorePath,
                                                                     "truststore_password", validTrustStorePassword);
@@ -345,7 +345,7 @@ public class AbstractEncryptionOptionsImpl extends TestBaseImpl
     }
 
     /* Provde the cluster cannot start with the configured options */
-    void assertCannotStartDueToConfigurationException(Cluster cluster)
+    protected void assertCannotStartDueToConfigurationException(Cluster cluster)
     {
         Throwable tr = null;
         try
